moves useWeb3Modal into web3 hooks
