+++ src/components/web3/hooks/index.ts
@@ -2,4 +2,5 @@ export * from './useContractSend';
 export * from './useETHGasPrice';
 export * from './useInitContracts';
 export * from './useIsDefaultChain';
+export * from './useWeb3Modal';
 export * from './useWeb3ModalManager';
