+++ src/pages/members/CreateMemberProposal.tsx
@@ -20,7 +20,7 @@ import {FormFieldErrors} from '../../util/enums';
 import {isEthAddressValid} from '../../util/validation';
 import {SHARES_ADDRESS} from '../../config';
 import {StoreState, MetaMaskRPCError} from '../../util/types';
-import {useWeb3Modal} from '../../components/web3/Web3ModalManager';
+import {useWeb3Modal} from '../../components/web3/hooks';
 import {Web3TxStatus} from '../../components/web3/types';
 import CycleMessage from '../../components/feedback/CycleMessage';
 import ErrorMessageWithDetails from '../../components/common/ErrorMessageWithDetails';
